% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LongLat.R
\name{LongLat}
\alias{LongLat}
\title{LongLat}
\usage{
LongLat(longitude, latitude)
}
\arguments{
\item{Longitude}{input (in degrees) vector to be converted to decimal minutes}

\item{Latitude}{input (in degrees) vector to be converted to decimal minutes}
}
\description{
This function converts location coordinate degrees to decimal minutes
}
